Backend-API/requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
